palm-api-apps/palm_analysis_action/requirements.txt
google-generativeai
python-dotenv

